split portion-with-blobs class on read and write parts
preparing for portion constructor object initiation
