+++ ydb/core/tx/columnshard/engines/changes/cleanup_tables.h
@@ -40,7 +40,7 @@ public:
     virtual ui32 GetWritePortionsCount() const override {
         return 0;
     }
-    virtual TPortionInfoWithBlobs* GetWritePortionInfo(const ui32 /*index*/) override {
+    virtual TWritePortionInfoWithBlobs* GetWritePortionInfo(const ui32 /*index*/) override {
         return nullptr;
     }
     virtual bool NeedWritePortion(const ui32 /*index*/) const override {
