+++ ydb/core/tx/columnshard/engines/changes/compaction/column_cursor.h
@@ -3,7 +3,6 @@
 #include <ydb/core/tx/columnshard/splitter/chunks.h>
 #include <ydb/core/tx/columnshard/engines/portions/column_record.h>
 #include <ydb/core/tx/columnshard/engines/scheme/column_features.h>
-#include <ydb/core/tx/columnshard/engines/portions/with_blobs.h>
 #include <contrib/libs/apache/arrow/cpp/src/arrow/record_batch.h>
 
 namespace NKikimr::NOlap::NCompaction {
