+++ ydb/core/tx/columnshard/engines/portions/base_with_blobs.cpp
@@ -1,0 +1,5 @@
+#include "base_with_blobs.h"
+
+namespace NKikimr::NOlap {
+
+}
