+++ ydb/core/tx/columnshard/engines/portions/base_with_blobs.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace NKikimr::NOlap {
+
+class TBasePortionInfoWithBlobs {
+public:
+};
+
+} // namespace NKikimr::NOlap
