+++ ydb/core/tx/columnshard/engines/portions/read_with_blobs.h
@@ -1,0 +1,69 @@
+#pragma once
+#include "base_with_blobs.h"
+#include "portion_info.h"
+#include <ydb/core/tx/columnshard/blob.h>
+#include <ydb/core/tx/columnshard/splitter/blob_info.h>
+#include <ydb/core/tx/columnshard/splitter/chunks.h>
+#include <ydb/core/tx/columnshard/engines/scheme/statistics/abstract/common.h>
+#include <ydb/core/tx/columnshard/engines/scheme/statistics/abstract/operator.h>
+
+#include <ydb/library/accessor/accessor.h>
+
+namespace NKikimr::NOlap {
+
+class TVersionedIndex;
+class TWritePortionInfoWithBlobs;
+
+class TReadPortionInfoWithBlobs: public TBasePortionInfoWithBlobs {
+private:
+    using TBlobChunks = std::map<TChunkAddress, std::shared_ptr<IPortionDataChunk>>;
+    YDB_READONLY_DEF(TBlobChunks, Chunks);
+    void RestoreChunk(const std::shared_ptr<IPortionDataChunk>& chunk);
+
+    TPortionInfo PortionInfo;
+    mutable std::optional<std::shared_ptr<arrow::RecordBatch>> CachedBatch;
+
+    explicit TReadPortionInfoWithBlobs(TPortionInfo&& portionInfo)
+        : PortionInfo(std::move(portionInfo)) {
+    }
+
+    explicit TReadPortionInfoWithBlobs(const TPortionInfo& portionInfo)
+        : PortionInfo(portionInfo) {
+    }
+
+    const TString& GetBlobByAddressVerified(const ui32 columnId, const ui32 chunkId) const;
+
+public:
+    static std::vector<TReadPortionInfoWithBlobs> RestorePortions(const std::vector<TPortionInfo>& portions, NBlobOperations::NRead::TCompositeReadBlobs& blobs,
+        const TVersionedIndex& tables);
+    static TReadPortionInfoWithBlobs RestorePortion(const TPortionInfo& portion, NBlobOperations::NRead::TCompositeReadBlobs& blobs,
+        const TIndexInfo& indexInfo);
+
+    std::shared_ptr<arrow::RecordBatch> GetBatch(const ISnapshotSchema::TPtr& data, const ISnapshotSchema& result, const std::set<std::string>& columnNames = {}) const;
+    static std::optional<TWritePortionInfoWithBlobs> SyncPortion(TReadPortionInfoWithBlobs&& source,
+        const ISnapshotSchema::TPtr& from, const ISnapshotSchema::TPtr& to, const TString& targetTier, const std::shared_ptr<IStoragesManager>& storages,
+        std::shared_ptr<NColumnShard::TSplitterCounters> counters);
+
+    std::vector<std::shared_ptr<IPortionDataChunk>> GetEntityChunks(const ui32 entityId) const;
+
+    bool ExtractColumnChunks(const ui32 columnId, std::vector<const TColumnRecord*>& records, std::vector<std::shared_ptr<IPortionDataChunk>>& chunks);
+
+    TString DebugString() const {
+        return TStringBuilder() << PortionInfo.DebugString() << ";";
+    }
+
+    const TPortionInfo& GetPortionInfo() const {
+        return PortionInfo;
+    }
+
+    TPortionInfo& GetPortionInfo() {
+        return PortionInfo;
+    }
+
+    friend IOutputStream& operator << (IOutputStream& out, const TReadPortionInfoWithBlobs& info) {
+        out << info.DebugString();
+        return out;
+    }
+};
+
+} // namespace NKikimr::NOlap
